update advanced options sections, add new checkbox
