+++ frontend/pages/admin/OrgSettingsPage/cards/Advanced/components/ActivityDataRetentionSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ActivityDataRetentionSection";
