+++ frontend/pages/admin/OrgSettingsPage/cards/Advanced/components/FeaturesSection/FeaturesSection.tsx
@@ -1,0 +1,103 @@
+import React from "react";
+import SettingsSection from "pages/admin/components/SettingsSection";
+import GitOpsModeTooltipWrapper from "components/GitOpsModeTooltipWrapper";
+import Checkbox from "components/forms/fields/Checkbox";
+
+import { IAdvancedSectionProps } from "../../Advanced";
+
+const FeaturesSection = ({
+  formData,
+  onInputChange,
+}: IAdvancedSectionProps) => {
+  const { disableLiveQuery, disableScripts, disableAIFeatures } = formData;
+
+  return (
+    <SettingsSection title="Features">
+      <GitOpsModeTooltipWrapper
+        position="left"
+        renderChildren={(disableChildren) => (
+          <Checkbox
+            disabled={disableChildren}
+            onChange={onInputChange}
+            name="disableLiveQuery"
+            value={disableLiveQuery}
+            parseTarget
+            labelTooltipContent={
+              !disableChildren && (
+                <>
+                  When enabled, disables the ability to run live reports <br />
+                  (ad hoc reports executed via the UI or fleetctl).{" "}
+                  <em>
+                    (Default: <strong>Off</strong>)
+                  </em>
+                </>
+              )
+            }
+          >
+            Disable live reports
+          </Checkbox>
+        )}
+      />
+
+      <GitOpsModeTooltipWrapper
+        position="left"
+        renderChildren={(disableChildren) => (
+          <Checkbox
+            disabled={disableChildren}
+            onChange={onInputChange}
+            name="disableScripts"
+            value={disableScripts}
+            parseTarget
+            labelTooltipContent={
+              !disableChildren && (
+                <>
+                  Disabling script execution will block access to run scripts.
+                  <br />
+                  Scripts may still be added and removed in the UI and API.
+                  <br />
+                  <em>
+                    (Default: <b>Off</b>)
+                  </em>
+                </>
+              )
+            }
+            helpText="Features that run scripts under-the-hood (e.g. software install, lock/wipe, script-only packages) will still be available."
+          >
+            Disable script execution features
+          </Checkbox>
+        )}
+      />
+      <GitOpsModeTooltipWrapper
+        position="left"
+        renderChildren={(disableChildren) => (
+          <Checkbox
+            disabled={disableChildren}
+            onChange={onInputChange}
+            name="disableAIFeatures"
+            value={disableAIFeatures}
+            parseTarget
+            labelTooltipContent={
+              !disableChildren && (
+                <>
+                  When enabled, disables AI features such as pre-filling forms
+                  <br />
+                  with descriptions generated by a large language model
+                  <br />
+                  (LLM).{" "}
+                  <em>
+                    (Default: <strong>Off</strong>)
+                  </em>
+                </>
+              )
+            }
+            helpText="If enabled, only policy queries (SQL) are sent to the LLM. Fleet doesn’t use this data to train models."
+          >
+            Disable generative AI features
+          </Checkbox>
+        )}
+      />
+    </SettingsSection>
+  );
+};
+
+export default FeaturesSection;
